Add files via upload
SQLite is different from other SQL databases because unlike most other SQL databases, SQLite does not have a separate server process. It reads and writes directly to ordinary disk files. SQLite performs better than SQL as It only loads the data which is needed, rather than reading the entire file and hold it in memory and Reading, writing operations are very fast for SQLite database; If you edit small parts, it only overwrite the parts of the file which was changed.
This project deals with Hotel Management System database in which we established a database connection with python and performed CRUD operations on it.